+++ src/helpers/prune-utils.ts
@@ -11,6 +11,8 @@ import { matchStackTrace } from './match-stack';
  * @param root object which should be pruned or logged
  * @param prunePaths array with string of space-separated property chains to remove
  * @param requiredPaths array with string of space-separated propertiy chains
+ * @param stack string which should be matched by stack trace
+ * @param nativeObjects reference to native objects, required for a trusted-prune-inbound-object to fix infinite loop
  * which must be all present for the pruning to occur
  * @returns true if prunning is required
  */
@@ -20,22 +22,25 @@ export function isPruningNeeded(
     prunePaths: string[],
     requiredPaths: string[],
     stack: string,
+    nativeObjects: any,
 ): boolean | undefined {
     if (!root) {
         return false;
     }
 
+    const { nativeStringify } = nativeObjects;
+
     let shouldProcess;
 
     // Only log hostname and matched JSON payload if only second argument is present
     if (prunePaths.length === 0 && requiredPaths.length > 0) {
-        const rootString = JSON.stringify(root);
+        const rootString = nativeStringify(root);
         const matchRegex = toRegExp(requiredPaths.join(''));
         const shouldLog = matchRegex.test(rootString);
         if (shouldLog) {
             logMessage(
                 source,
-                `${window.location.hostname}\n${JSON.stringify(root, null, 2)}\nStack trace:\n${new Error().stack}`,
+                `${window.location.hostname}\n${nativeStringify(root, null, 2)}\nStack trace:\n${new Error().stack}`,
                 true,
             );
             if (root && typeof root === 'object') {
@@ -95,6 +100,8 @@ export function isPruningNeeded(
  * @param root object which should be pruned or logged
  * @param prunePaths array with string of space-separated properties to remove
  * @param requiredPaths array with string of space-separated properties
+ * @param stack string which should be matched by stack trace
+ * @param nativeObjects reference to native objects, required for a trusted-prune-inbound-object to fix infinite loop
  * which must be all present for the pruning to occur
  * @returns pruned root
  */
@@ -104,11 +111,13 @@ export const jsonPruner = (
     prunePaths: string[],
     requiredPaths: string[],
     stack: string,
+    nativeObjects: any,
 ): ArbitraryObject => {
+    const { nativeStringify } = nativeObjects;
     if (prunePaths.length === 0 && requiredPaths.length === 0) {
         logMessage(
             source,
-            `${window.location.hostname}\n${JSON.stringify(root, null, 2)}\nStack trace:\n${new Error().stack}`,
+            `${window.location.hostname}\n${nativeStringify(root, null, 2)}\nStack trace:\n${new Error().stack}`,
             true,
         );
         if (root && typeof root === 'object') {
@@ -118,7 +127,7 @@ export const jsonPruner = (
     }
 
     try {
-        if (isPruningNeeded(source, root, prunePaths, requiredPaths, stack) === false) {
+        if (isPruningNeeded(source, root, prunePaths, requiredPaths, stack, nativeObjects) === false) {
             return root;
         }
 
@@ -138,4 +147,21 @@ export const jsonPruner = (
     }
 
     return root;
+};
+
+/**
+ * Checks if props is a string and returns array of properties
+ * or empty array if props is not a string
+ *
+ * @param props string of space-separated properties or undefined
+ * @returns array of properties or empty array if props is not a string
+ */
+export const getPrunePath = (props: unknown) => {
+    const validPropsString = typeof props === 'string'
+        && props !== undefined
+        && props !== '';
+
+    return validPropsString
+        ? props.split(/ +/)
+        : [];
 };